+++ backend/cwa_lib/pages/query_documents.py
@@ -1,6 +1,6 @@
 from typing import Sequence
 from sqlalchemy import select
-from common.enums.gvdbs_cfg_json import GVDBsCfgJSON
+from common.features.gvdbs_retr_params import GVDBsRetrParams, GVDBsDefRetrParams
 from common.sql_db_async import AsyncSession
 from common.sql_models import User, GroupContexts, GroupLLMs, GroupVDBs
 from cwa_lib.pydantic_schemas.doc_tasks import (
@@ -10,7 +10,7 @@ from cwa_lib.pydantic_schemas.doc_tasks import (
     DocTasksOptionsGroupContextsRow, 
     DocTasksOptionsGroupLLMsRow, 
     DocTasksOptionsGroupVDBsRow,
-    GVDBsCfgDefaults
+    DocTasksOptionsGVDBsDefRetrParams
 )
 from cwa_lib.sql_tables.doc_tasks import DocTasksTable
 from cwa_lib.sql_tables.api_settings import ApiSettingsTable
@@ -37,7 +37,7 @@ class QueryDocumentsPage:
                 group_id=self.user.group_id, 
                 user_id=self.user.user_id, 
                 gvdbs_id=payload.gvdbs_id,
-                gvdbs_cfg_json=GVDBsCfgJSON.from_dict(payload.gvdbs_cfg_json).as_dict(),
+                gvdbs_cfg_json=GVDBsRetrParams.from_dict(payload.gvdbs_cfg_json).as_dict(),
                 gllms_id=payload.gllms_id,
                 gc_id=payload.gc_id,
                 short_name=payload.short_name, 
@@ -49,7 +49,7 @@ class QueryDocumentsPage:
                 doc_task_id=payload.doc_task_id,
                 user_group_id=self.user.group_id,
                 gvdbs_id=payload.gvdbs_id,
-                gvdbs_cfg_json=GVDBsCfgJSON.from_dict(payload.gvdbs_cfg_json).as_dict(),
+                gvdbs_cfg_json=GVDBsRetrParams.from_dict(payload.gvdbs_cfg_json).as_dict(),
                 gllms_id=payload.gllms_id,
                 gc_id=payload.gc_id,
                 short_name=payload.short_name, 
@@ -80,10 +80,11 @@ class QueryDocumentsOptions:
         result = await self.session.execute(select(GroupVDBs).where(where_clause).order_by(GroupVDBs.gvdbs_id))
         return result.scalars().all()
     
-    async def _get_gvdbs_cfg_defaults(self) -> GVDBsCfgDefaults:
-        gvdbs_cfg_json = (await ApiSettingsTable(self.session).select_by_names(['gvdbs_cfg_json']))['gvdbs_cfg_json']
-        gvdbs_cfg_obj = GVDBsCfgJSON.from_dict(gvdbs_cfg_json)        
-        return GVDBsCfgDefaults(search_type=gvdbs_cfg_obj.search_type, search_kwargs=gvdbs_cfg_obj.search_kwargs)
+    async def _get_gvdbs_def_retr_params(self) -> DocTasksOptionsGVDBsDefRetrParams:
+        gvdbs_def_retr_params = (await ApiSettingsTable(self.session).select_by_names(['gvdbs_def_retr_params']))['gvdbs_def_retr_params']
+        gvdbs_def_retr_params_obj = GVDBsDefRetrParams.from_dict(gvdbs_def_retr_params)
+        # TODO: value should be taken from `group_vdbs` or `api_groups` or `api_settings`
+        return DocTasksOptionsGVDBsDefRetrParams(**gvdbs_def_retr_params_obj.__dict__)
 
     async def get_options(self) -> DocTaskOptionsResult:
         """
@@ -92,12 +93,11 @@ class QueryDocumentsOptions:
             2) `group_llms`
             3) `group_vdbs`
         Also:
-        `gvdbs_cfg_defaults`: default values for `doc_tasks.gvdbs_cfg_json`.
+        `gvdbs_def_retr_params`: default values for `doc_tasks.gvdbs_def_retr_params`.
         """
-        
         return DocTaskOptionsResult(
             group_contexts=await self._from__group_contexts(),
             group_llms=await self._from__group_llms(),
             group_vdbs=await self._from__group_vdbs(),
-            gvdbs_cfg_defaults=await self._get_gvdbs_cfg_defaults(),
+            gvdbs_def_retr_params=await self._get_gvdbs_def_retr_params(),
         )