+++ backend/cwa_lib/pages/server_status.py
@@ -1,4 +1,3 @@
-from time import time
 from traceback import format_exc
 from common import log
 from common.watchdogs import get_outdated_status