+++ frontend/src/pages/QueryDocuments/models/docTasksQuery.ts
@@ -1,9 +1,9 @@
-import type { DocTasksGVDBsCfgState } from "../../../components/GVDBsCfg/stores";
+import type { DocTasksGVDBsCfgState } from "../../../components/GVDBsRetrParams/types";
 
 export type DocTasksQuery = {
   doc_task_id: number | null;
   gvdbs_id: number;
-  gvdbs_cfg_json?: DocTasksGVDBsCfgState;
+  gvdbs_cfg_json: DocTasksGVDBsCfgState;
   gllms_id: number;
   gc_id: number;
   short_name: string;